refactor package and file names
