+++ src/main/java/org/opensearch/flowframework/common/ThrowingSupplier.java
@@ -6,7 +6,7 @@
  * this file be licensed under the Apache-2.0 license or a
  * compatible open source license.
  */
-package org.opensearch.flowframework.function;
+package org.opensearch.flowframework.common;
 
 /**
  * A supplier that can throw checked exception
